Use the Double primative for JSON Schema "number" type (#111)
Problem
=======
The JSON Schema stipulates that [a "number" type can be an integer, floating point, or exponential
notation](https://json-schema.org/understanding-json-schema/reference/numeric).
Currently, the "number" type is treated the same as an integer when
instantiating `fromJSONSchema`. In those cases, providing a float value
of e.g. `2.5` will fail because it can't be converted into a BigInt.

Solution
========
I changed the primitive to Double for "number" types when creating a schema from JSON.

Co-authored-by: Wil Wade <wil.wade@unfinished.com>